Android: moves chrome/ to use long for JNI
BUG=317523

Review URL: https://codereview.chromium.org/67573003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@236278 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/android/java/src/org/chromium/chrome/browser/autofill/AutofillDialogControllerAndroid.java
@@ -20,7 +20,7 @@ public class AutofillDialogControllerAndroid {
     private static AutofillDialogFactory sDialogFactory;
     private static boolean sAllowInsecureDialogsForTesting = false;
 
-    private int mNativeDelegate;  // could be 0 after onDestroy().
+    private long mNativeDelegate;  // could be 0 after onDestroy().
     private AutofillDialog mDialog;
 
     /**
@@ -109,7 +109,7 @@ public class AutofillDialogControllerAndroid {
     }
 
     private AutofillDialogControllerAndroid(
-            final int nativeAutofillDialogControllerAndroid,
+            final long nativeAutofillDialogControllerAndroid,
             final WindowAndroid windowAndroid,
             final boolean requestFullBillingAddress, final boolean requestShippingAddress,
             final boolean requestPhoneNumbers,
@@ -160,7 +160,7 @@ public class AutofillDialogControllerAndroid {
 
     @CalledByNative
     private static AutofillDialogControllerAndroid create(
-            final int nativeAutofillDialogControllerAndroid,
+            final long nativeAutofillDialogControllerAndroid,
             final WindowAndroid windowAndroid,
             final boolean requestFullBillingAddress, final boolean requestShippingAddress,
             final boolean requestPhoneNumbers,
@@ -199,8 +199,8 @@ public class AutofillDialogControllerAndroid {
 
     // Calls from Java to C++ AutofillDialogControllerAndroid:
 
-    private native void nativeDialogCancel(int nativeAutofillDialogControllerAndroid);
-    private native void nativeDialogContinue(int nativeAutofillDialogControllerAndroid,
+    private native void nativeDialogCancel(long nativeAutofillDialogControllerAndroid);
+    private native void nativeDialogContinue(long nativeAutofillDialogControllerAndroid,
             Object fullWallet,
             boolean lastUsedChoiceIsAutofill, String lastUsedAccountName,
             String guidLastUsedBilling, String guidLastUsedShipping, String guidLastUsedCard);
