+++ chrome/android/java/src/org/chromium/chrome/browser/profiles/Profile.java
@@ -12,9 +12,9 @@ import org.chromium.base.CalledByNative;
  */
 public class Profile {
 
-    private int mNativeProfileAndroid;
+    private long mNativeProfileAndroid;
 
-    private Profile(int nativeProfileAndroid) {
+    private Profile(long nativeProfileAndroid) {
         mNativeProfileAndroid = nativeProfileAndroid;
     }
 
@@ -23,7 +23,7 @@ public class Profile {
     }
 
     @CalledByNative
-    private static Profile create(int nativeProfileAndroid) {
+    private static Profile create(long nativeProfileAndroid) {
         return new Profile(nativeProfileAndroid);
     }
 
@@ -33,7 +33,7 @@ public class Profile {
     }
 
     @CalledByNative
-    private int getNativePointer() {
+    private long getNativePointer() {
         return mNativeProfileAndroid;
     }
