+++ chrome/browser/prerender/external_prerender_handler_android.h
@@ -31,7 +31,7 @@ class ExternalPrerenderHandlerAndroid {
   bool AddPrerender(JNIEnv* env,
                     jobject obj,
                     jobject profile,
-                    jint web_content_ptr,
+                    jlong web_content_ptr,
                     jstring url,
                     jstring referrer,
                     jint width,
